+++ src/test/kotlin/org/rooftop/order/infra/transaction/Fixture.kt
@@ -1,6 +1,7 @@
-package org.rooftop.order.app
+package org.rooftop.order.infra.transaction
 
 import org.rooftop.order.domain.OrderState
+import org.rooftop.order.infra.transaction.UndoOrder
 
 fun undoOrder(
     id: Long = 1L,
